// core/app/core/src/lib/containers/confirmation-list-modal/models/confirmation-list-modal.model.ts
/**
 * SuiteCRM is a customer relationship management program developed by SuiteCRM Ltd.
 * Copyright (C) 2026 SuiteCRM Ltd.
 *
 * This program is free software; you can redistribute it and/or modify it under
 * the terms of the GNU Affero General Public License version 3 as published by the
 * Free Software Foundation with the addition of the following permission added
 * to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
 * IN WHICH THE COPYRIGHT IS OWNED BY SUITECRM, SUITECRM DISCLAIMS THE
 * WARRANTY OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
 *
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
 * details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 *
 * In accordance with Section 7(b) of the GNU Affero General Public License
 * version 3, these Appropriate Legal Notices must retain the display of the
 * "Supercharged by SuiteCRM" logo. If the display of the logos is not reasonably
 * feasible for technical reasons, the Appropriate Legal Notices must display
 * the words "Supercharged by SuiteCRM".
 */

import {ActionData, ActionHandler} from '../../../common/actions/action.model';
import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
import {RecordListModalStore} from '../../record-list-modal/store/record-list-modal/record-list-modal.store';

export interface ConfirmationListFieldConfig {
    name: string;
    label?: string;
    labelKey?: string;
    type?: string;
    link?: boolean;
    [key: string]: any;
}

export interface ConfirmationListAction {
    key: string;
    labelKey: string;
    asyncProcess?: boolean;
    klass?: string[];
    params?: { [key: string]: any };
}

export interface ConfirmationListActionData extends ActionData {
    module: string;
    store: RecordListModalStore;
    activeModal: NgbActiveModal;
    onProceed: Function;
    onClose: Function;
}

export abstract class ConfirmationListActionHandler extends ActionHandler<ConfirmationListActionData> {

    abstract run(data: ConfirmationListActionData): void;

    shouldDisplay(data: ConfirmationListActionData): boolean {
        return true;
    }
}

export interface ConfirmationListActionsAdapterConfig {
    module: string;
    store: RecordListModalStore;
    activeModal: NgbActiveModal;
    onProceed: Function;
    onClose: Function;
    actions: ConfirmationListAction[];
}